Remove underscore.js dependency
