+++ lib/activedirectory.js
@@ -2,7 +2,6 @@ var events = require('events');
 var util = require('util');
 var ldap = require('ldapjs');
 var async = require('async');
-var _ = require('underscore');
 var bunyan = require('bunyan');
 
 var Group = require('./models/group');
@@ -279,45 +278,6 @@ ActiveDirectory.prototype.emit = function emit(event, data) {
     events.EventEmitter.prototype.emit.call(this, event, data);
   }
 };
-
-/**
- * Factory to create the LDAP client object.
- *
- * @private
- * @param {String} url The url to use when creating the LDAP client.
- * @param {object} opts The optional LDAP client options.
- */
-function createClient(url, opts) {
-  // Attempt to get Url from this instance.
-  url = url || this.url || (this.opts || {}).url || (opts || {}).url;
-  if (! url) {
-    throw 'No url specified for ActiveDirectory client.';
-  }
-  log.trace('createClient(%s)', url);
-
-  var opts = getLdapClientOpts(_.defaults({}, { url: url }, opts, this.opts));
-  log.debug('Creating ldapjs client for %s. Opts: %j', opts.url, _.omit(opts, 'url', 'bindDN', 'bindCredentials'));
-  var client = ldap.createClient(opts);
-  return(client);
-}
-
-/**
- * From the list of options, retrieves the ldapjs client specific options.
- *
- * @param {Object} opts The opts to parse.
- * @returns {Object} The ldapjs opts.
- */
-function getLdapClientOpts(opts) {
-  return(_.pick(opts || {},
-    // Client
-    'url',
-    'host', 'port', 'secure', 'tlsOptions',
-    'socketPath', 'log', 'timeout', 'idleTimeout',
-    'reconnect', 'queue', 'queueSize', 'queueTimeout',
-    'queueDisable', 'bindDN', 'bindCredentials',
-    'maxConnections'
-  ));
-}
 
 /**
  * Gets all of the groups that the specified distinguishedName (DN) belongs to.
@@ -984,7 +944,7 @@ ActiveDirectory.prototype.authenticate = function authenticate(username, passwor
     return callback(err, false);
   }
 
-  const client = createClient.call(this);
+  const client = utils.createClient(null, null, this);
   client.on('error', (err) => {
     // only used on socket connection failure since it doesn't invoke bind cb
     this.emit('error', err);
@@ -1038,7 +998,7 @@ ActiveDirectory.prototype.getRootDSE = function getRootDSE(url, attributes, call
   }
   log.trace('getRootDSE(%s,%j)', _url, _attributes);
 
-  const client = createClient.call(this, _url);
+  const client = utils.createClient.call(this, _url, this);
   client.on('error', (err) => {
     // only needed for bind errors
     if (err.errno !== 'ECONNRESET') { // we don't care about ECONNRESET
