+++ lib/components/search.js
@@ -2,57 +2,11 @@
 
 const url = require('url');
 const ldap = require('ldapjs');
+const utils = require('./utilities');
 
 let ad;
 let log;
 let parseRangeAttributes;
-
-/**
- * Factory to create the LDAP client object.
- *
- * @private
- * @param {String} url The url to use when creating the LDAP client.
- * @param {object} opts The optional LDAP client options.
- */
-function createClient(url, opts) {
-  // Attempt to get Url from this instance.
-  const _url = url || ad.url || ad.opts.url;
-  if (!_url) {
-    throw 'No url specified for ActiveDirectory client.';
-  }
-  log.trace('createClient(%s)', _url);
-
-  const ldapOpts = getLdapClientOpts(Object.assign({ url: _url }, opts));
-  log.debug('Creating ldapjs client for %s. Opts: %j', opts.url, ldapOpts);
-  return ldap.createClient(opts);
-}
-
-/**
- * From the list of options, retrieves the ldapjs client specific options.
- *
- * @private
- * @param {Object} opts The opts to parse.
- * @returns {Object} The ldapjs opts.
- */
-function getLdapClientOpts(opts) {
-  const clientOpts = [
-    'url',
-    'host', 'port', 'secure', 'tlsOptions',
-    'socketPath', 'log', 'timeout', 'idleTimeout',
-    'reconnect', 'queue', 'queueSize', 'queueTimeout',
-    'queueDisable', 'bindDN', 'bindCredentials',
-    'maxConnections'
-  ];
-
-  const options = {};
-  clientOpts.forEach((opt) => {
-    if (opts.hasOwnProperty(opt)) {
-      options[opt] = opts[opt];
-    }
-  });
-
-  return options;
-}
 
 function isReferralAllowed(referralUri) {
   if (ad.defaultReferrals.enabled === false || !referralUri) {
@@ -94,7 +48,7 @@ function Searcher(opts) { // jshint -W071
   this.baseDN = opts.baseDN;
   this.callback = opts.callback;
   this.opts = opts;
-  this.ldapOpts = getLdapClientOpts(this.opts);
+  this.ldapOpts = utils.getLdapClientOpts(this.opts);
   this.query = this.opts.opts;
 
   this.results = [];
@@ -102,7 +56,7 @@ function Searcher(opts) { // jshint -W071
   this.pendingRangeRetrievals = 0;
   this.searchComplete = false;
 
-  this.client = createClient(ad.url || ad.opts.url, this.ldapOpts);
+  this.client = utils.createClient(ad.url || ad.opts.url, this.ldapOpts, ad);
 
   this.controls = this.opts.controls || [];
