+++ lib/components/utilities.js
@@ -1,7 +1,29 @@
 'use strict';
 
+const ldap = require('ldapjs');
+
 // this module consists of various utility functions that are used
 // throughout the ActiveDirectory code
+
+/**
+ * Factory to create the LDAP client object.
+ *
+ * @private
+ * @param {String} url The url to use when creating the LDAP client.
+ * @param {object} opts The optional LDAP client options.
+ * @param {ActiveDirectory} An instance of {@link ActiveDirectory} to pull
+ * default values from.
+ */
+function createClient(url, opts, ad) {
+  // Attempt to get Url from this instance.
+  const _url = url || ad.url || ad.opts.url;
+  if (!_url) {
+    throw 'No url specified for ActiveDirectory client.';
+  }
+
+  const ldapOpts = getLdapClientOpts(Object.assign({ url: _url }, opts));
+  return ldap.createClient(ldapOpts);
+}
 
 /**
  * Gets a properly formatted LDAP compound filter. This is a very simple
@@ -40,6 +62,33 @@ function getGroupQueryFilter(groupName) {
       parseDistinguishedName(groupName) + '))';
   }
   return `(&(objectCategory=Group)(cn=${groupName}))`;
+}
+
+/**
+ * From the list of options, retrieves the ldapjs client specific options.
+ *
+ * @private
+ * @param {Object} opts The opts to parse.
+ * @returns {Object} The ldapjs opts.
+ */
+function getLdapClientOpts(opts) {
+  const clientOpts = [
+    'url',
+    'host', 'port', 'secure', 'tlsOptions',
+    'socketPath', 'log', 'timeout', 'idleTimeout',
+    'reconnect', 'queue', 'queueSize', 'queueTimeout',
+    'queueDisable', 'bindDN', 'bindCredentials',
+    'maxConnections'
+  ];
+
+  const options = {};
+  clientOpts.forEach((opt) => {
+    if (opts.hasOwnProperty(opt)) {
+      options[opt] = opts[opt];
+    }
+  });
+
+  return options;
 }
 
 /**
@@ -293,8 +342,10 @@ function truncateLogOutput(output, maxLength) {
 }
 
 module.exports = {
+  createClient,
   getCompoundFilter,
   getGroupQueryFilter,
+  getLdapClientOpts,
   getRequiredLdapAttributesForGroup,
   getRequiredLdapAttributesForUser,
   getUserQueryFilter,
