+++ package.json
@@ -20,8 +20,7 @@
   "dependencies": {
     "async": ">= 0.1.22",
     "bunyan": ">= 1.3.5",
-    "ldapjs": ">= 0.7.1",
-    "underscore": ">= 1.4.3"
+    "ldapjs": ">= 0.7.1"
   },
   "scripts": {
     "test": "mocha --ui bdd test",
